lsp: Decouple virtual file system
